Optimized code and added better error handling

+++ operations.py
@@ -10,10 +10,10 @@ conn = psycopg2.connect(database = DATABASE_NAME,
                         password = PASSWORD,
                         port = PORT)
 
-def getAllStudents():
-    # Open a cursor to perform database operations
-    cur = conn.cursor()
+# Open a cursor to perform database operations
+cur = conn.cursor()
 
+def getAllStudents():
     # Execute a command: select all data from students table
     cur.execute("""SELECT * FROM students
                    ORDER BY student_id ASC""")
@@ -34,62 +34,75 @@ def getAllStudents():
     conn.close()
 
 def addStudent(first_name, last_name, email, enrollment_date):
-    # Open a cursor to perform database operations
-    cur = conn.cursor()
-
-    # Execute a command: insert data into students table
     try:
+        # Execute a command: insert data into students table
         cur.execute(f"""INSERT INTO students (first_name, last_name, email, enrollment_date) 
                         VALUES('{first_name}','{last_name}','{email}','{enrollment_date}')""")
         # Make the changes to the database persistent
         conn.commit()
-        print('Success!')
+        
+        print('Student added successfully!')
+        print("Here's the updated table: \n")
+        getAllStudents()
 
     except Exception as error:
         print("ERROR:", error)
-    
-    # Close cursor and communication with the database
-    cur.close()
-    conn.close()
+        # Close cursor and communication with the database
+        cur.close()
+        conn.close()
 
 def updateStudentEmail(student_id, new_email):
-    # Open a cursor to perform database operations
-    cur = conn.cursor()
-
-    # Execute a command: modify a students email
     try:
+        #Validate student_id
+        cur.execute(f"""SELECT * FROM students
+                        WHERE student_id = {student_id}""")
+
+        if cur.fetchall() == []:
+            raise Exception("student_id does not exist in table")
+        
+        # Execute a command: modify a students email
         cur.execute(f"""UPDATE students 
                         SET email = '{new_email}' 
                         WHERE student_id = {student_id}""")
         # Make the changes to the database persistent
         conn.commit()
-        print('Success!')
+
+        print('Student email updated successfully!')
+        print("Here's the updated table: \n")
+        getAllStudents()
+
 
     except Exception as error:
         print("ERROR:", error)
+        # Close cursor and communication with the database
+        cur.close()
+        conn.close()
     
-    # Close cursor and communication with the database
-    cur.close()
-    conn.close()
 
 def deleteStudent(student_id):
-    # Open a cursor to perform database operations
-    cur = conn.cursor()
-
-    # Execute a command: delete student from database
     try:
+         #Validate student_id
+        cur.execute(f"""SELECT * FROM students
+                        WHERE student_id = {student_id}""")
+        if cur.fetchall() == []:
+            raise Exception("student_id does not exist in table")
+        
+        # Execute a command: delete student from database
         cur.execute(f"""DELETE from students 
                         WHERE student_id = {student_id}""")
         # Make the changes to the database persistent
         conn.commit()
-        print('Success!')
+
+        print('Student deleted successfully!')
+        print("Here's the updated table: \n")
+        getAllStudents()
 
     except Exception as error:
         print("ERROR:", error)
-    
-    # Close cursor and communication with the database
-    cur.close()
-    conn.close()
+         # Close cursor and communication with the database
+        cur.close()
+        conn.close()
+
     
 if __name__ == '__main__':
     args = sys.argv
